Codebase transformation
Added missing script files + moved everything to their namespaces based on the folder name + fixed references inside Unity Editor
