+++ Unity Project/Assets/Code/Scripts/SampleScene/FinishScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3d78ca152983c254bbc101a4f6334db3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
