+++ Unity Project/Assets/Code/Scripts/SampleScene/FlashLight.cs.meta
