+++ Unity Project/Assets/Code/Scripts/SampleScene/PickUpObjects.cs.meta
