+++ Unity Project/Assets/Code/Scripts/SampleScene/Player/MoveCamera.cs
@@ -1,0 +1,14 @@
+using UnityEngine;
+
+namespace Code.Scripts.SampleScene.Player
+{
+    public class MoveCamera : MonoBehaviour
+    {
+        public Transform cameraPosition;
+
+        public void Update()
+        {
+            transform.position = cameraPosition.position;
+        }
+    }
+}
