+++ Unity Project/Assets/Code/Scripts/SampleScene/Player/MoveCamera.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c0fcd5144c55af499915d352913e077
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
