+++ Unity Project/Assets/Code/Scripts/SampleScene/Player/PickUpScript.cs.meta
