+++ Unity Project/Assets/Code/Scripts/SampleScene/Player/PlayerCam.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 08500199cac716a44ac504b468f00fc5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
