+++ Unity Project/Assets/Code/Scripts/SampleScene/Player/PlayerMovment.cs
@@ -1,0 +1,83 @@
+using UnityEngine;
+
+namespace Code.Scripts.SampleScene.Player
+{
+    public class PlayerMovment : MonoBehaviour
+    {
+        [Header("Movment")]
+        public float moveSpeed;
+
+        public float groundDrag;
+
+        [Header("Ground Check")]
+        public float playerHeight;
+        public LayerMask whatIsGround;
+        bool grounded;
+
+        public Transform orientation;
+
+        float horizontalInput;
+        float verticalInput;
+
+        Vector3 moveDirection;
+
+        Rigidbody rb;
+
+        private void Start()
+        {
+            rb = GetComponent<Rigidbody>();
+            rb.freezeRotation = true;
+        }
+
+        private void Update()
+        {
+            MyInput();
+        }
+
+        private void LateUpdate()
+        {
+            // ground check
+            grounded = Physics.Raycast(transform.position, Vector3.down, playerHeight * 0.5f + 0.2f, whatIsGround);
+
+            MyInput();
+            SpeedControl();
+
+            // handle drag
+            if (grounded)
+                rb.drag = groundDrag;
+            else
+                rb.drag = 0;
+        }
+
+        private void FixedUpdate()
+        {
+            MovePlayer();
+        }
+
+        private void MyInput()
+        {
+            horizontalInput = Input.GetAxisRaw("Horizontal");
+            verticalInput = Input.GetAxisRaw("Vertical");
+        }
+
+        private void MovePlayer()
+        {
+            // calculate movment direction
+            moveDirection = orientation.forward * verticalInput + orientation.right * horizontalInput;
+
+            rb.AddForce(moveDirection.normalized * moveSpeed * 10f, ForceMode.Force);
+        }
+
+        private void SpeedControl()
+        {
+            Vector3 flatVel = new Vector3(rb.velocity.x, 0f, rb.velocity.z);
+
+            // limit velocity if needed
+            if (flatVel.magnitude > moveSpeed)
+            {
+                Vector3 limitedVel = flatVel.normalized * moveSpeed;
+                rb.velocity = new Vector3(limitedVel.x, rb.velocity.y, limitedVel.z);
+            }
+        }
+    }
+}
