+++ Unity Project/Assets/Code/Scripts/SampleScene/Player/PlayerMovment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1271a259d1ae19a408f108ee3d4f7b6c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
