+++ Unity Project/Assets/Code/Scripts/SampleScene/Player/Tags.cs
@@ -1,0 +1,19 @@
+using UnityEngine;
+
+namespace Code.Scripts.SampleScene.Player
+{
+    public class Tags : MonoBehaviour
+    {
+        // Start is called before the first frame update
+        void Start()
+        {
+        
+        }
+
+        // Update is called once per frame
+        void Update()
+        {
+        
+        }
+    }
+}
