+++ Unity Project/Assets/Level/Scenes/SampleScene.unity
@@ -122,7 +122,7 @@ NavMeshSettings:
     preserveTilesOutsideBounds: 0
     debug:
       m_Flags: 0
-  m_NavMeshData: {fileID: 23800000, guid: 370ea627625a2af4eaab41893263ac54, type: 2}
+  m_NavMeshData: {fileID: 23800000, guid: 67bbad405d4dd9b42abe9d3fc62103a8, type: 2}
 --- !u!1 &9700934
 GameObject:
   m_ObjectHideFlags: 0
@@ -1675,7 +1675,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b7f9931e9ac663d4e8216dba48f4efc4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  itemButton: {fileID: 0}
+  itemButton: {fileID: 8903918739483906253, guid: 740abb9fd06755f4cbcc56374248be63, type: 3}
 --- !u!1 &182141967
 GameObject:
   m_ObjectHideFlags: 0
@@ -3521,8 +3521,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 6573c40ac446be14287f8991bec14eae, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  pauseMenuUI: {fileID: 0}
-  InventoryScript: {fileID: 0}
+  pauseMenuUI: {fileID: 1841126147}
+  InventoryScript: {fileID: 1874861933}
 --- !u!1 &319986424
 GameObject:
   m_ObjectHideFlags: 0
@@ -4742,13 +4742,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a707d76d1ae1af8429e1f04e61f1d726, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Player: {fileID: 1874861933}
-  PauseMenu: {fileID: 308034127}
-  InventoryScript: {fileID: 1874861933}
-  target: {fileID: 0}
-  agent: {fileID: 0}
-  lookRadius: 100
-  RestartMenu: {fileID: 2020305496}
 --- !u!54 &417766099
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -5140,14 +5133,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 451595782}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 49ceda6ad5e723e4285fedd656489a14, type: 3}
+  m_Script: {fileID: 11500000, guid: 08500199cac716a44ac504b468f00fc5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   sensX: 400
   sensY: 400
   orientation: {fileID: 285932653}
-  xRotation: 0
-  yRotation: 0
 --- !u!81 &451595785
 AudioListener:
   m_ObjectHideFlags: 0
@@ -7527,7 +7518,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b7f9931e9ac663d4e8216dba48f4efc4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  itemButton: {fileID: 0}
+  itemButton: {fileID: 8903918739483906253, guid: c19d95863271e8e459e6b8fe886b31f6, type: 3}
 --- !u!1 &680629008
 GameObject:
   m_ObjectHideFlags: 0
@@ -9921,7 +9912,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b7f9931e9ac663d4e8216dba48f4efc4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  itemButton: {fileID: 0}
+  itemButton: {fileID: 8903918739483906253, guid: 9fb4cf710c78921499aafec102696aff, type: 3}
 --- !u!1 &913494947
 GameObject:
   m_ObjectHideFlags: 0
@@ -10366,13 +10357,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a707d76d1ae1af8429e1f04e61f1d726, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Player: {fileID: 1874861933}
-  PauseMenu: {fileID: 308034127}
-  InventoryScript: {fileID: 1874861933}
-  target: {fileID: 0}
-  agent: {fileID: 0}
-  lookRadius: 500
-  RestartMenu: {fileID: 2020305496}
 --- !u!54 &954472117
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -12883,7 +12867,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1134824191}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 604b1e6581bd8eb4a98a6c28d3e41d94, type: 3}
+  m_Script: {fileID: 11500000, guid: 4876614255e6ac649a8ac876a115adcb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   player: {fileID: 1874861933}
@@ -14123,13 +14107,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: a707d76d1ae1af8429e1f04e61f1d726, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  Player: {fileID: 1874861933}
-  PauseMenu: {fileID: 308034127}
-  InventoryScript: {fileID: 1874861933}
-  target: {fileID: 0}
-  agent: {fileID: 0}
-  lookRadius: 54.8
-  RestartMenu: {fileID: 2020305496}
 --- !u!54 &1228816652
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -18543,7 +18520,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1678131467}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: da3a8cf8380ffa0488336447750d219f, type: 3}
+  m_Script: {fileID: 11500000, guid: 5c0fcd5144c55af499915d352913e077, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   cameraPosition: {fileID: 1876077929}
@@ -20778,7 +20755,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b7f9931e9ac663d4e8216dba48f4efc4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  itemButton: {fileID: 0}
+  itemButton: {fileID: 3238933211843698238, guid: a59938f6c5508854a93bc384d3c57b87, type: 3}
 --- !u!1 &1817487686
 GameObject:
   m_ObjectHideFlags: 0
@@ -21762,20 +21739,16 @@ MonoBehaviour:
   m_GameObject: {fileID: 1874861933}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 69f4f43e49e23a941bf879b23a4f1d21, type: 3}
+  m_Script: {fileID: 11500000, guid: 1271a259d1ae19a408f108ee3d4f7b6c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   moveSpeed: 55
   groundDrag: 5
   playerHeight: 2
   whatIsGround:
+    serializedVersion: 2
     m_Bits: 8
-  grounded: 0
   orientation: {fileID: 285932653}
-  horizontalInput: 0
-  verticalInput: 0
-  moveDirection: {x: 0, y: 0, z: 0}
-  rb: {fileID: 0}
 --- !u!54 &1874861936
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -21868,13 +21841,18 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 09037609187eb244f8094b909cf4100e, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  isFull: 
-  slots: []
-  HandisFull: 
-  Handslots: []
+  isFull: 000000
+  slots:
+  - {fileID: 1159948154}
+  - {fileID: 1781091999}
+  - {fileID: 1163106828}
+  HandisFull: 0000
+  Handslots:
+  - {fileID: 1488334629}
+  - {fileID: 646100969}
   scripts: []
-  Inventory: {fileID: 0}
-  pauseMenu: {fileID: 0}
+  Inventory: {fileID: 450578447}
+  pauseMenu: {fileID: 308034127}
 --- !u!114 &1874861941
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -21889,8 +21867,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   selectableTag: Selectable
   HandTag: Handtag
-  PressEToPickUp: {fileID: 0}
-  rayLength: 0
+  PressEToPickUp: {fileID: 1829788513}
+  rayLength: 3
 --- !u!1 &1876077928
 GameObject:
   m_ObjectHideFlags: 0
@@ -22407,14 +22385,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 1915035369}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a707d76d1ae1af8429e1f04e61f1d726, type: 3}
+  m_Script: {fileID: 11500000, guid: 4b654123c240bd14f8a3168efd33a54a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   Player: {fileID: 1874861933}
   PauseMenu: {fileID: 308034127}
   InventoryScript: {fileID: 1874861933}
-  target: {fileID: 0}
-  agent: {fileID: 0}
   lookRadius: 100
   RestartMenu: {fileID: 2020305496}
 --- !u!1 &1916186808
@@ -22946,7 +22922,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1981712561}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: fe3843cb0613fca4fbe924b61c23bb84, type: 3}
+  m_Script: {fileID: 11500000, guid: 3d78ca152983c254bbc101a4f6334db3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   Player: {fileID: 1463110187}
@@ -23453,7 +23429,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 9d5f5132fd11c6c4dbe6e9d769e110a4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  restartMenu: {fileID: 0}
+  restartMenu: {fileID: 2020305496}
 --- !u!1 &2021547987
 GameObject:
   m_ObjectHideFlags: 0
@@ -24362,7 +24338,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b7f9931e9ac663d4e8216dba48f4efc4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  itemButton: {fileID: 0}
+  itemButton: {fileID: 8273284563218733026, guid: edd97d57cd31aa0408de8dd2f39d37a8, type: 3}
 --- !u!1 &2100428177
 GameObject:
   m_ObjectHideFlags: 0
