+++ Unity Project/Assets/Level/Scenes/SampleScene/NavMesh.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 67bbad405d4dd9b42abe9d3fc62103a8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 23800000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
