+++ Unity Project/ProjectSettings/EditorBuildSettings.asset
@@ -6,12 +6,12 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/MainMenu.unity
+    path: Assets/Level/Scenes/MainMenu.unity
     guid: e982be3ce772a924dbe34e680cf6dd2a
   - enabled: 1
-    path: Assets/Scenes/SampleScene.unity
+    path: Assets/Level/Scenes/SampleScene.unity
     guid: 9fc0d4010bbf28b4594072e72b8655ab
   - enabled: 1
-    path: Assets/Scenes/Level2.unity
+    path: Assets/Level/Scenes/Level2.unity
     guid: ef8fcca215963da45815812db02b1b65
   m_configObjects: {}
